tests: Temporarily disable spec tests
